+++ assets/javascripts/kitten/components/dropdowns/phone-dropdown.js
@@ -27,6 +27,20 @@ class PhoneDropdown extends React.Component {
     this.handlePositionUpdate()
 
     emitter.on('dropdown:opening:trigger', this.handleOtherDropdownsOpening)
+
+    if (this.props.closeEvents) {
+      this.props.closeEvents.forEach((ev) => {
+        window.addEventListener(ev, this.handleOtherDropdownsOpening)
+      })
+    }
+  }
+
+  componentWillUnmount() {
+    if (this.props.closeEvents) {
+      this.props.closeEvents.forEach((ev) => {
+        window.removeEventListener(ev, this.handleOtherDropdownsOpening)
+      })
+    }
   }
 
   // Component methods.
@@ -184,6 +198,7 @@ PhoneDropdown.propTypes = {
   positionedWithBorder: React.PropTypes.bool,
   positionedOn: React.PropTypes.string,
   refreshEvents: React.PropTypes.array,
+  closeEvents: React.PropTypes.array,
   onPositionUpdate: React.PropTypes.func,
 }